code clean up, api for timeedit
